Remove ADDA and SUBA cases from InstructionDecoder
The `case` statements for `OpHandlerID.ADDA` and `OpHandlerID.SUBA`
were removed from the `switch` block in `InstructionDecoder.cs`.
These cases previously included an assertion to check
`inst.SourceAddrMode` and a call to `ReadImmediateOperandData`
to retrieve operand data.

This change backs out the previous addition of this code that broke the instruction execution.