Removing ext1.HasValue checks as now the OpCodeDecoder and InstructionDecoder handle the addressing mode checks once per instruction rather than for each use of an instruction.  Small performance benefit, but mostly just to clean up the code.
As per the previous commit, SingeStepTests now have 56 passing and 71 failing with the fix of the (d8,An,Xn) address mode.

As for the purpose of these changes, the Mac SE/HD ROM is still getting stuck with multiple address errors/HALT with a jump to PC $00010001 (an odd address) after loading the SCSI OS image and displaying the Welcome to Macintosh banner than then the (empty) menu bar.